+++ lib/src/_impl/xplat/_worker_monitor.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import '../../tokens/cancellation_token.dart';
 import '../../worker/worker_pool.dart';
 import '../../worker/worker_request.dart';
@@ -5,19 +7,21 @@ import '../../worker/worker_service.dart';
 import '../../worker/worker_task.dart';
 import '_cancellation_token_ref.dart';
 
-/// Each platform worker will instantiate a [WorkerMonitor] responsible for handling cancellation requests.
-/// Worker tasks in Squadron may be cancelled in two ways: with a [CancellationToken], giving worker services
-/// the chance to handle cancellation requests gracefully, or without a [CancellationToken] via
-/// [WorkerPool.cancel] or [Task.cancel].
+/// Each platform worker will instantiate a [WorkerMonitor] responsible for
+/// handling cancellation requests. Worker tasks in Squadron may be cancelled in
+/// two ways: with a [CancellationToken], giving worker services the chance to
+/// handle cancellation requests gracefully, or without a [CancellationToken]
+/// via [WorkerPool.cancel] or [Task.cancel].
 class WorkerMonitor {
-  /// Constructs a new [WorkerMonitor]. The [_terminate] callback will be called by
-  /// [WorkerMonitor.terminate]. This callback should contain code to shutdown the
-  /// platform worker and release resources.
+  /// Constructs a new [WorkerMonitor]. The [_terminate] callback will be called
+  /// by [WorkerMonitor.terminate]. This callback should contain code to
+  /// shutdown the platform worker and release resources.
   WorkerMonitor(this._terminate);
 
   final SquadronCallback _terminate;
   bool _terminationRequested = false;
   int _executing = 0;
+  ServiceInstaller? _installer;
 
   final _cancelTokens = <String, CancellationTokenReference>{};
 
@@ -75,7 +79,29 @@ class WorkerMonitor {
     }
     _executing--;
     if (_terminationRequested && _executing == 0) {
-      _terminate();
+      _shutdown();
+    }
+  }
+
+  /// Installs the service if it implements [ServiceInstaller].
+  FutureOr<void> install(WorkerService service) {
+    if (service is ServiceInstaller) {
+      final installer = service as ServiceInstaller;
+      _installer = installer;
+      return installer.install();
+    }
+  }
+
+  /// Installs the service if it implements [ServiceInstaller].
+  FutureOr<void> _uninstall(ServiceInstaller? installer) {
+    if (installer != null) {
+      try {
+        var res = installer.uninstall();
+        if (res is Future<void>) {
+          res = res.onError((error, stackTrace) => {/* swallow excepions */});
+        }
+        return res;
+      } catch (ex) {/* swallow exceptions*/}
     }
   }
 
@@ -83,9 +109,18 @@ class WorkerMonitor {
   /// the worker as terminating and termination will be effective when all pending executions are completed.
   void terminate() {
     if (_executing == 0) {
-      _terminate();
+      _shutdown();
     } else {
       _terminationRequested = true;
+    }
+  }
+
+  void _shutdown() {
+    final uninstall = _uninstall(_installer);
+    if (uninstall is Future) {
+      uninstall.whenComplete(_terminate);
+    } else {
+      _terminate();
     }
   }
 }