+++ test/sample_vm_workers/installable_worker.dart
@@ -1,0 +1,10 @@
+import 'package:squadron/squadron.dart';
+
+import '../worker_services/installable_service.dart';
+
+void start(List command) => run(
+    (startRequest) => InstallableService(
+        throwOnInstall: startRequest.args[0],
+        throwOnUninstall: startRequest.args[1]),
+    command,
+    ParentSquadronLogger());