+++ test/sample_vm_workers/worker_entry_points.dart
@@ -1,6 +1,7 @@
 import '../worker_services/worker_entry_points.dart' show EntryPoints;
 
 import 'cache_worker.dart' as cache;
+import 'installable_worker.dart' as installable;
 import 'issues_worker.dart' as issues;
 import 'local_client_worker.dart' as local_client;
 import 'prime_worker.dart' as prime;
@@ -10,6 +11,7 @@ const platform = 'NATIVE';
 
 void setEntryPoints() {
   EntryPoints.cache = cache.start;
+  EntryPoints.installable = installable.start;
   EntryPoints.issues = issues.start;
   EntryPoints.local = local_client.start;
   EntryPoints.prime = prime.start;